Add option to allow namespaces in non-terminals
The flag allows non-terminals in EBNF grammars to contain `.` and `/`.
The non-terminal must not start with a `/` or `.` to avoid clashing
with `/` and `.` for `:ord` rules and rule ends respectively.

Other than that there are no restrictions to keep it simple. If a user
uses the flag and then forgets to pad `:ord` `/`s with spaces, then the
parser will parse `a/b` as a single namespaced non-terminal and an
error will be reported if that non-terminal can't be found, which
should be enough to help resolve the issue.

Similarly, there is no restriction on the number and order of `.` and
`/` that can show up in the non-terminal, only that the non-terminal
must not start with either. Clojure and ClojureScript may handle these
differently when they are being turned into keywords, but that is left
to the user.

Fixes Engelberg/instaparse#211